Tests do not ModifyPluginSettings anymore
