+++ layers/GraphQLAPIForWP/phpunit-packages/graphql-api-for-wp/tests/Integration/DisableSchemaModulesOnPersistedQueryEndpointTestOnCustomAdminEndpointsFixtureEndpointWebserverRequestTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace PHPUnitForGraphQLAPI\GraphQLAPI\Integration;
 
-class DisableSchemaModulesOnPersistedQueryEndpointTestOnCustomAdminEndpointsModifyPluginSettingsFixtureEndpointWebserverRequestTest extends AbstractDisableSchemaModulesOnPrivateEndpointsModifyPluginSettingsFixtureEndpointWebserverRequestTest
+class DisableSchemaModulesOnPersistedQueryEndpointTestOnCustomAdminEndpointsFixtureEndpointWebserverRequestTest extends AbstractDisableSchemaModulesOnPrivateEndpointsFixtureEndpointWebserverRequestTest
 {
     protected function getEndpoint(): string
     {
