--- layers/GraphQLAPIForWP/phpunit-packages/graphql-api-for-wp/tests/Integration/DisableSchemaModulesOnPrivateEndpointNoChangeAdminEndpointsModifyPluginSettingsFixtureEndpointWebserverRequestTestTrait.php
@@ -1,23 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace PHPUnitForGraphQLAPI\GraphQLAPI\Integration;
-
-trait DisableSchemaModulesOnPrivateEndpointNoChangeAdminEndpointsModifyPluginSettingsFixtureEndpointWebserverRequestTestTrait
-{
-    protected function getResponseFixtureFolder(): string
-    {
-        return __DIR__ . '/fixture-disable-schema-modules-on-private-endpoints-no-change';
-    }
-
-    /**
-     * @param array<string,mixed> $providerItems
-     * @return array<string,mixed>
-     */
-    protected function customizeProviderEndpointEntries(array $providerItems): array
-    {
-        $providerItems['type-introspection'][1] = $providerItems['type-introspection:0'][1];
-        return $providerItems;
-    }
-}
