+++ layers/GraphQLAPIForWP/phpunit-packages/graphql-api-for-wp/tests/Integration/DisableSchemaModulesOnPrivateEndpointTestOnBlockEditorAdminEndpointsFixtureEndpointWebserverRequestTest.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PHPUnitForGraphQLAPI\GraphQLAPI\Integration;
+
+use GraphQLAPI\GraphQLAPI\Constants\AdminGraphQLEndpointGroups;
+
+class DisableSchemaModulesOnPrivateEndpointTestOnBlockEditorAdminEndpointsFixtureEndpointWebserverRequestTest extends AbstractDisableSchemaModulesOnPrivateEndpointTestOnCustomAdminEndpointsFixtureEndpointWebserverRequestTest
+{
+    use DisableSchemaModulesOnPrivateEndpointNoChangeAdminEndpointsFixtureEndpointWebserverRequestTestTrait;
+
+    protected function getAdminEndpointGroup(): string
+    {
+        return AdminGraphQLEndpointGroups::BLOCK_EDITOR;
+    }
+}
