+++ layers/GraphQLAPIForWP/phpunit-packages/graphql-api-for-wp/tests/Integration/fixture-disable-schema-modules-on-private-endpoints-no-change/type-introspection.json
@@ -1,1 +1,9 @@
-{}
+{
+    "data": {
+        "__schema": {
+            "userType": {
+                "name": "User"
+            }
+        }
+    }
+}
