--- layers/GraphQLAPIForWP/phpunit-packages/graphql-api-for-wp/tests/Integration/fixture-disable-schema-modules-on-private-endpoints-no-change/type-introspection:0.json
@@ -1,9 +1,0 @@
-{
-    "data": {
-        "__schema": {
-            "userType": {
-                "name": "User"
-            }
-        }
-    }
-}
