--- layers/GraphQLAPIForWP/phpunit-packages/graphql-api-for-wp/tests/Integration/fixture-disable-schema-modules-on-private-endpoints-persisted-query/user-account-query-placeholder:0.json
@@ -1,26 +1,0 @@
-{
-    "errors": [
-        {
-            "message": "There is no field 'user' on type 'QueryRoot'",
-            "locations": [
-                {
-                    "line": 5,
-                    "column": 3
-                }
-            ],
-            "extensions": {
-                "path": [
-                    "user(by: {id: $userID}) { ... }",
-                    "query GetUserAccountData($userID: ID!, $limit: Int) { ... }"
-                ],
-                "type": "QueryRoot",
-                "field": "user(by: {id: $userID}) { ... }",
-                "code": "gql@5.3.1",
-                "specifiedBy": "https:\/\/spec.graphql.org\/draft\/#sec-Field-Selections"
-            }
-        }
-    ],
-    "data": {
-        "user": null
-    }
-}
