--- layers/GraphQLAPIForWP/phpunit-packages/graphql-api-for-wp/tests/Integration/fixture-disable-schema-modules-on-private-endpoints-public-endpoint/type-introspection:0.json
@@ -1,7 +1,0 @@
-{
-    "data": {
-        "__schema": {
-            "userType": null
-        }
-    }
-}
